+++ kafka_test_teardown.sh
@@ -4,6 +4,7 @@ set -ex
 
 echo "Unregistering test topics"
 build/kafka/bin/kafka-topics.sh --zookeeper localhost:2181 --delete --topic 'logstash_integration_.*'
+build/kafka/bin/kafka-topics.sh --zookeeper localhost:2181 --delete --topic 'topic_avro.*'
 
 echo "Stopping Kafka broker"
 build/kafka/bin/kafka-server-stop.sh